requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Django==5.0
exceptiongroup==1.2.0
fonttools==4.47.0
frozendict==2.3.10
h11==0.14.0
html5lib==1.1
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.4
matplotlib==3.8.2
multitasking==0.0.11
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
peewee==3.17.0
Pillow==10.1.0
plotly==5.18.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tenacity==8.2.3
textblob==0.17.1
threadpoolctl==3.2.0
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
webdriver-manager==4.0.1
webencodings==0.5.1
wsproto==1.2.0
yfinance==0.2.33